requirements-dev.txt
pytest>=6.2.4
python-dotenv>=0.20.0
cassandra-driver>=3.24.0,<4
